python提高课件/day13-django1/day13_spider/requirements.txt
certifi==2021.5.30
chardet==3.0.4
idna==2.5
pycryptodome==3.10.1
PyExecJS==1.5.1
PyMySQL==1.0.2
redis==3.5.3
requests==2.18.0
six==1.16.0
urllib3==1.21.1
